BUG: bsplines: fixed issue 14525
bspline, quadratic and cubic gave zero outputs for integer dtypes.
Fixed through casting to float in the already present asarray.
This fixes it but does not find the logic flaw.

+++ scipy/signal/bsplines.py
@@ -182,7 +182,7 @@ def bspline(x, n):
     True
 
     """
-    ax = -abs(asarray(x))
+    ax = -abs(asarray(x,dtype=float))
     # number of pieces on the left-side is (n+1)/2
     funclist, condfuncs = _bspline_piecefunctions(n)
     condlist = [func(ax) for func in condfuncs]
@@ -280,7 +280,7 @@ def cubic(x):
     True
 
     """
-    ax = abs(asarray(x))
+    ax = abs(asarray(x,dtype=float))
     res = zeros_like(ax)
     cond1 = less(ax, 1)
     if cond1.any():
@@ -332,7 +332,7 @@ def quadratic(x):
     True
 
     """
-    ax = abs(asarray(x))
+    ax = abs(asarray(x,dtype=float))
     res = zeros_like(ax)
     cond1 = less(ax, 0.5)
     if cond1.any():
